# go-server-server/go/api_product.go
/*
 * Swagger Petstore - OpenAPI 3.0
 *
 * This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach! You can now help us improve the API whether it's by making changes to the definition itself or to the code. That way, with time, we can improve the API in general, and expose some of the new features in OAS3.
 *
 * API version: 1.0.11
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

import (
	"encoding/json"
	"net/http"
	"strings"

	"github.com/gorilla/mux"
)

// ProductApiController binds http requests to an api service and writes the service results to the http response
type ProductApiController struct {
	service ProductApiServicer
	errorHandler ErrorHandler
}

// ProductApiOption for how the controller is set up.
type ProductApiOption func(*ProductApiController)

// WithProductApiErrorHandler inject ErrorHandler into controller
func WithProductApiErrorHandler(h ErrorHandler) ProductApiOption {
	return func(c *ProductApiController) {
		c.errorHandler = h
	}
}

// NewProductApiController creates a default api controller
func NewProductApiController(s ProductApiServicer, opts ...ProductApiOption) Router {
	controller := &ProductApiController{
		service:      s,
		errorHandler: DefaultErrorHandler,
	}

	for _, opt := range opts {
		opt(controller)
	}

	return controller
}

// Routes returns all the api routes for the ProductApiController
func (c *ProductApiController) Routes() Routes {
	return Routes{ 
		{
			"AddProduct",
			strings.ToUpper("Post"),
			"/api/v3/product",
			c.AddProduct,
		},
		{
			"DeleteProduct",
			strings.ToUpper("Delete"),
			"/api/v3/product/{productId}",
			c.DeleteProduct,
		},
		{
			"GetProduct",
			strings.ToUpper("Get"),
			"/api/v3/product/{productId}",
			c.GetProduct,
		},
		{
			"GetProductList",
			strings.ToUpper("Get"),
			"/api/v3/product",
			c.GetProductList,
		},
		{
			"UpdateProduct",
			strings.ToUpper("Put"),
			"/api/v3/product",
			c.UpdateProduct,
		},
	}
}

// AddProduct - Add a new Product to the store
func (c *ProductApiController) AddProduct(w http.ResponseWriter, r *http.Request) {
	productParam := Product{}
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&productParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertProductRequired(productParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.AddProduct(r.Context(), productParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}

// DeleteProduct - Deletes a Product
func (c *ProductApiController) DeleteProduct(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	productIdParam, err := parseInt64Parameter(params["productId"], true)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}

	result, err := c.service.DeleteProduct(r.Context(), productIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}

// GetProduct - Find Product by ID
func (c *ProductApiController) GetProduct(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	productIdParam, err := parseInt64Parameter(params["productId"], true)
	if err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}

	result, err := c.service.GetProduct(r.Context(), productIdParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}

// GetProductList - Get list of all products
func (c *ProductApiController) GetProductList(w http.ResponseWriter, r *http.Request) {
	result, err := c.service.GetProductList(r.Context())
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}

// UpdateProduct - Update an existing Product
func (c *ProductApiController) UpdateProduct(w http.ResponseWriter, r *http.Request) {
	productParam := Product{}
	d := json.NewDecoder(r.Body)
	d.DisallowUnknownFields()
	if err := d.Decode(&productParam); err != nil {
		c.errorHandler(w, r, &ParsingError{Err: err}, nil)
		return
	}
	if err := AssertProductRequired(productParam); err != nil {
		c.errorHandler(w, r, err, nil)
		return
	}
	result, err := c.service.UpdateProduct(r.Context(), productParam)
	// If an error occurred, encode the error with the status code
	if err != nil {
		c.errorHandler(w, r, err, &result)
		return
	}
	// If no error, encode the body and the result code
	EncodeJSONResponse(result.Body, &result.Code, w)

}