# go-server-server/go/model_product.go
/*
 * Swagger Petstore - OpenAPI 3.0
 *
 * This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach! You can now help us improve the API whether it's by making changes to the definition itself or to the code. That way, with time, we can improve the API in general, and expose some of the new features in OAS3.
 *
 * API version: 1.0.11
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Product struct {

	Id int64 `json:"id,omitempty"`

	Name string `json:"name,omitempty"`

	Quantity int64 `json:"quantity,omitempty"`

	Price int64 `json:"price,omitempty"`
}

// AssertProductRequired checks if the required fields are not zero-ed
func AssertProductRequired(obj Product) error {
	return nil
}

// AssertRecurseProductRequired recursively checks if required fields are not zero-ed in a nested slice.
// Accepts only nested slice of Product (e.g. [][]Product), otherwise ErrTypeAssertionError is thrown.
func AssertRecurseProductRequired(objSlice interface{}) error {
	return AssertRecurseInterfaceRequired(objSlice, func(obj interface{}) error {
		aProduct, ok := obj.(Product)
		if !ok {
			return ErrTypeAssertionError
		}
		return AssertProductRequired(aProduct)
	})
}